Allow GET method on ssh-keys and pgp-keys
The /keys/ssh-keys and /keys/pgp-keys POST endpoints used to
add keys to an account do not redirect the user if an error
occurs during validation. Because of this, trying to reload
the page after an error occurs can cause a Method Not Allowed
error if the browser changes the request to a GET request.
This commit fixes this by allowing the GET method on the endpoints,
and simply rendering the keys.html template in this scenario.

Signed-off-by: Benjamin Lowry <ben@ben.gmbh>