metasrht-createuser: Allow reading password from stdin
Currently we use getpass() which will read from /dev/tty which is annoying
to automate.

This commit adds a new option '-s' which will make the password be read
directly from stdin instead. For example this lets you do:

    tr -c -d '!-~' < /dev/urandom | head -c 64 | tee \
      | metasrht-createuser -s -t admin user user@example.com

To create a new admin user with a random password 64 characters in length.