Move account registration into GQL
This moves the account registration process into GraphQL and updates
meta.sr.ht to be a thin shim on top of it. This also adds a database
constraint to prevent duplicate emails - thankfully there was only one.

This change also fixes a couple of lingering bugs: attempting to sign up
with a PGP key already in the system now shows an error, and a few
places where key_id was still used in the UI were replaced with
fingerprint_hex.