Handle unexpected responses during API discovery
Some services (at least man.sr.ht) do not yet have a GraphQL API
backend. The discovery will still attempt to contact the origin from the
config file. In the case of man.sr.ht, if no default wiki is set up, it
will respond for _any_ path with a 200 and a default HTML page. This
breaks the API discovery and makes meta.sr.ht crash.

This fixes this by not only checking for a 200 return code but also
attempting to parse the JSON before accepting the response and
gracefully continue if it cannot be parsed.