api: use EncryptionKey() to test PGP key validity
The `EncryptionKey()` function "returns the best candidate Key for
encrypting a message to the given Entity" [1]. In doing so, it performs
various sanity checks on the key [2].

This works as a pretty good proxy for key validation, as encrypting an
email to someone is the primary use case. The validation is much more
thorough than what `CanSign()` does.

[1]:
https://godocs.io/github.com/ProtonMail/go-crypto/openpgp#Entity.EncryptionKey
[2]:
https://github.com/ProtonMail/go-crypto/blob/master/openpgp/keys.go#L121-L163