Start tracking PGP key expiration in database
It would be great to be able to inform users if their uploaded PGP key
is expired or about to expire. The expiration date of a key is somewhat
complicated to compute and doesn't change unless the key material is
updated. Hence, compute it during key upload and store it in the
database.

Upon migration, all existing keys will have the default value of NULL,
meaning no expiration. This is not correct, but also not a problem. New
keys will have their expiration date set correctly. A one-off job can be
run for existing keys to compute their expiration date. The added
directory `contrib/key-expiration` contains a small Go program that can
be used for this purpose.

With this change, the expiration date is only stored, it is not yet used
for anything.

Fixes: https://todo.sr.ht/~sircmpwn/meta.sr.ht/194