api: try to find PGP key's primary fingerprint
PGP key material can contain hard-to-make-sense-of data. We currently
rely on go-crypto to find a suitable encryption key, the only thing we
care about (for sending emails). The downside of that approach is that
we currently display the encryption key's fingerprint, which more often
than not is actually a sub-key. Hence, the fingerprint will not match
what the user expects.

The current approach was mainly to err on the side of caution and leave
the interpretation of the key material to go-crypto. But since we have
this as fallback, there is no harm in at least trying to get the key's
main fingerprint, which will work as long as the key material has a
baseline reasonable structure.

This approach still has the same robustness, but will be less confusing
for users who supply sane key material.