contrib/key-expiration: handle failure cases
This script must keep going in face of all errors. All errors are now
handled and printed to stdout, the loop then continues. The only error
that is still treated as fatal is a database error.

The approach to finding the expiration date was also too simplistic, as
the `EncryptionKey()` function fails if the key is already expired.
Hence add an attempt to find the expiration date manually in this case.
This is a tricky issue, so the situation where `EncryptionKey()` fails
but no expiration date can be determined is also logged.