api: SendEmailNotification takes formatted message
This is the execution of the plan put forth in the discussion of [1]
(see [2] for full thread). Make the (currently unused) internal mutation
`SendEmailNotification` take a readily formatted message in IMF [3],
with the added restriction that it must not contain any recipients.

The handling consists of adding the recipient, encrypting the body if
appropriate, and sending it out.

This interface should be generic enough to allow for all current use
cases of sending email to be centralized in this API call.

[1]: https://lists.sr.ht/~sircmpwn/sr.ht-dev/patches/31069
[2]: https://lists.sr.ht/~sircmpwn/sr.ht-dev/%3C20220413134751.78596-1-ch%40bitfehler.net%3E
[3]: https://www.rfc-editor.org/rfc/rfc5322