Allow private-use scheme for OAuth 2.0 redirect URIs
RFC 8252 describes best OAuth 2.0 practices for native apps. Section 7
explains that authorization servers MUST offer support for three redirect
URI options, one of which is "Private-Use URI Scheme Redirection"
described in section 7.1. It allows clients to set the redirect URI to e.g.

    com.example.app:/oauth2redirect/example-provider

And leverages the platform mechanism to associate an app to an URI
scheme.