Accept and check redirect_uri parameter on OAuth 2.0 token endpoint
Some OAuth 2.0 clients will always send it. The RFC says it's optional,
so clients don't expect it to fail. Instead of forcing them to add
special handling for sr.ht, tolerate the parameter as long as it matches
the one configured in the registered OAuth client.

Same as [1] but for the token endpoint.

[1]: https://lists.sr.ht/~sircmpwn/sr.ht-dev/patches/35260