api: make sendEmailNotification recipient explicit
Having sendEmailNotification take a full message as argument is very
powerful, but restricting it to only send mail to the currently
logged-in user will in fact not get us very far. Examples of mails that
should be sent (potentially encrypted) that are not triggered by a
logged-in user include the password reset email, billing emails
triggered by metasrht-daily, and possibly more in other services.

This commit extends the API to include the username of the desired
recipient. This will not add much overhead to the callers and more
importantly, still make sure to keep any PGP-related code in meta.sr.ht.