api: add endpoint to download invoices
This commit adds an endpoint - though not a GraphQL one - to the API to
generate PDF invoices for download. The endpoint is compatible with the
current one (requires a POST with the same parameters), but the frontend
is not yet using it, so that it can receive more widespread testing
before going "live".

The PDF generator takes the same parameters as the Python implementation
and it's output is reasonably close. An example can be found at [1]. It
includes the smallest, sanely-licensed monospace font that a quick
search on my system turned up. PDF comes with default fonts, but they
don't support UTF-8, and IMO messing with encodings/code pages by far
outweighs the inconveniences of embedding a font or messing with
font-config to find one on the system (e.g. on my system, the standard
"monospace" font maps to an OTF font, which is not supported).

[1]: https://bitfehler.net/invoice_test.pdf

Related: https://todo.sr.ht/~sircmpwn/meta.sr.ht/183