api: send OAuth authorization email to user
Currently, the notification email for having authorized a third party
OAuth client is not sent to the acting user, but instead to the owner of
the client. This commit fixes that.

The OAuth2 third party client authorization notification email is
triggered by a request from the third party client to the
`/oauth2/authorize` enpoint (see RFC 6749 section 4.1.1 [1]). This
request has no user authentication in the sense of the auth middleware.
Instead, it happens when processing the OAuth parameters, after all
middlewares. Hence, `auth.ForContext()` returns a context that contains
the OAuth client and its owner, but not the user performing the action.
Due to that, the email currently gets send to the owner.

To rectify, provide a `sendSecurityNotificationTo()` function that does
not use the current auth context at all and use it for this specific use
case. The regular `sendSecurityNotification()` just becomes a thin
wrapper around that function providing all values from the current auth
context.

While at it, tweak it to also get the PGP key from the context, because
that's what was used on all call sites anyway.

[1]: https://www.rfc-editor.org/rfc/rfc6749#section-4.1.1