api: add generic `sendEmail` mutation
This mutation acts as a replacement for both `sendEmailNotification` and
`sendEmailExternal`. It takes as address and a message as argument.
The message will always be signed with the site key. If the address is
that of a registered user, it will also be encrypted according to the
user's privacy settings.

Both `sendEmailNotification` and `sendEmailExternal` will be removed as
soon as all callers have been ported to use `sendEmail`.