metasrht-daily: add delay to billing attempts
There's an issue with Stripe which is causing a lot of unusual behavior
around failed charges. Stripe support has been utterly useless in
diagnosing this but I'm going to slow down billing on the unfounded
assumption that one of their broken heuristics has to do with the
frequency of charge attempts.